data-collection/src/extractors/requirements.txt
# Legal RAG Extraction System - Phase 3
# Requirements file for metadata extraction

# Core dependencies
pydantic>=2.0.0,<3.0.0          # Data validation and configuration
pydantic-settings>=2.0.0        # Settings management (Pydantic v2)
python-dotenv>=1.0.0            # Environment variable management
PyYAML>=6.0                     # Pattern file parsing

# PDF extraction
PyPDF2>=3.0.0                   # PDF reading
pdfplumber>=0.10.0              # PDF text extraction
pdfminer.six>=20221105          # Advanced PDF parsing
python-magic>=0.4.27            # File type detection

# OCR (optional but recommended)
pytesseract>=0.3.10             # Tesseract OCR wrapper
pdf2image>=1.16.0               # PDF to image conversion
Pillow>=10.0.0                  # Image processing

# HTML extraction
beautifulsoup4>=4.12.0          # HTML parsing
lxml>=4.9.0                     # XML/HTML parser
html5lib>=1.1                   # HTML5 parser

# Text processing
unicodedata2>=15.0.0            # Unicode handling
python-dateutil>=2.8.0          # Date parsing
regex>=2023.0.0                 # Advanced regex

# Machine learning
scikit-learn>=1.3.0             # TF-IDF and classification
numpy>=1.24.0                   # Numerical computing
scipy>=1.11.0                   # Scientific computing

# Database (for Phase 2 integration)
SQLAlchemy>=2.0.0               # Database ORM
psycopg2-binary>=2.9.0          # PostgreSQL adapter

# Utilities
tqdm>=4.65.0                    # Progress bars
colorlog>=6.7.0                 # Colored logging

# Development dependencies (optional)
# Uncomment for development
# pytest>=7.4.0                  # Testing framework
# pytest-cov>=4.1.0              # Coverage reporting
# black>=23.0.0                  # Code formatting
# flake8>=6.0.0                  # Linting
# mypy>=1.5.0                    # Type checking
# ipython>=8.14.0                # Interactive shell
# jupyter>=1.0.0                 # Jupyter notebooks

# System dependencies (install via apt/yum):
# - tesseract-ocr (for OCR)
# - poppler-utils (for pdf2image)
# - libmagic1 (for python-magic)
# - postgresql (for Phase 2 integration)
#
# Ubuntu/Debian:
#   sudo apt-get install tesseract-ocr poppler-utils libmagic1 postgresql
#
# macOS:
#   brew install tesseract poppler libmagic postgresql
